refactor(OMN-250): extract shared OmniJS-bridge test harness (run-analytics-script)
Deduplicates the vm.runInNewContext boilerplate flagged in /code-review of
PRs #209/#213; the helper is byte-identical on both branches so the add/add
merges cleanly in either order.
